Log out text is now red
