+++ lib/features/profile/profile_view.dart
@@ -4,6 +4,8 @@ import 'package:book_adapter/controller/firebase_controller.dart';
 import 'package:book_adapter/features/profile/change_password_view.dart';
 import 'package:book_adapter/features/profile/edit_profile_view.dart';
 import 'package:book_adapter/features/profile/profile_view_controller.dart';
+import 'package:book_adapter/features/profile/widgets/change_password_button.dart';
+import 'package:book_adapter/features/profile/widgets/log_out_button.dart';
 import 'package:book_adapter/features/profile/widgets/profile_widget.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -27,7 +29,7 @@ class ProfileView extends ConsumerWidget {
       appBar: AppBar(
         title: const Text(title),
         actions: const [
-          _LogOutButton(),
+          LogOutButton(),
         ],
       ),
       body: SingleChildScrollView(
@@ -56,65 +58,6 @@ class ProfileView extends ConsumerWidget {
           ),
         ),
       )
-    );
-  }
-}
-
-class _LogOutButton extends ConsumerWidget {
-  const _LogOutButton({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final viewController = ref.watch(profileViewController.notifier);
-    return IconButton(
-      key: const ValueKey('signOut'),
-      icon: const Icon(Icons.logout),
-      tooltip: 'Sign out',
-      onPressed: () {
-        // Log out the user
-        showDialog(
-          context: context,
-          builder: (BuildContext context) {
-            return AlertDialog(
-              title: const Text('Logout'),
-              content: const Text('Are you sure you want to log out?'),
-              actions: [
-                TextButton(
-                  onPressed: () {
-                    Navigator.of(context).pop();
-                  },
-                  child: const Text('CANCEL'),
-                ),
-                TextButton(
-                  onPressed: () {
-                    Navigator.of(context).popUntil(ModalRoute.withName('/'));
-                    viewController.signOut();
-                  },
-                  child: const Text('LOGOUT'),
-                ),
-              ],
-            );
-          },
-        );
-      },
-    );
-  }
-}
-
-class ChangePasswordButton extends StatelessWidget {
-  const ChangePasswordButton({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      child: const Text('Change Password'),
-      onPressed: () {
-        Navigator.restorablePushNamed(context, ChangePasswordView.routeName);
-      },
     );
   }
 }
